+++ CHANGELOG.md
@@ -4,15 +4,17 @@
 
 ## [Unreleased]
 
+<<<<<<< HEAD
 ## [0.2.1] - 2026-08-07
 
 ### Added 
 
-- Added audio cue for when AudioBug initially starts up. 
+- Added audio cue for when AudioBug initially starts up.
+- Added startup announcement.
 
-## [0.2.0] - 2026-08-05
 
-### Added
+## [0.2] - 2026-08-05
+>>>>>>> components/AudioBugStartupNotification
 
 - Added error/warning cycling for when multiple errors or warnings exist on the same line.